+++ packages/validation/tests/validate.spec.ts
@@ -1,47 +1,53 @@
-import type { Description } from '@sonata-api/types'
 import assert from 'assert'
 import { unwrapEither, isLeft, isRight } from '@sonata-api/common'
 import { validate } from '../dist'
 
-const description: Omit<Description, '$id'> = {
-  properties: {
-    name: {
-      type: 'string'
-    },
-    age: {
-      type: 'number'
-    }
-  }
-}
+import {
+  plainCandidate,
+  plainDescription,
+  deepCandidate,
+  deepDescription
+
+} from './fixtures'
 
 describe('Validate', () => {
-  it('validates with no errors', async () => {
-    const candidate = {
-      name: 'Terry',
-      age: 50
-    }
-
-    const validationEither = await validate(candidate, description)
+  it('validates plain object', () => {
+    const validationEither = validate(plainCandidate, plainDescription)
     assert(isRight(validationEither))
-    assert(JSON.stringify(candidate) === JSON.stringify(unwrapEither(validationEither)))
+    assert(JSON.stringify(plainCandidate) === JSON.stringify(unwrapEither(validationEither)))
   })
 
-  it('returns error on invalid type', async () => {
-    const candidate = {
-      name: 'Terry',
-      age: '50'
-    }
+  it('returns error on plain object', () => {
+    const candidate = Object.assign({}, plainCandidate)
+    candidate.age = '0' as any
 
-    const validationEither = await validate(candidate, description)
+    const validationEither = validate(candidate, plainDescription)
 
     assert(isLeft(validationEither))
     const error = unwrapEither(validationEither)
-
     assert(error.code === 'INVALID_PROPERTIES')
-    // assert(error.errors?.age?.type === 'unmatching')
-    // assert(error.errors?.age?.details.expected === 'number')
-    // assert(error.errors?.age?.details.got === 'string')
   })
 
+  it('validates deep object', () => {
+    const validationEither = validate(deepCandidate, deepDescription, {
+      recurse: true
+    })
+
+    assert(isRight(validationEither))
+    assert(JSON.stringify(deepCandidate) === JSON.stringify(unwrapEither(validationEither)))
+  })
+
+  it('returns error on deep object', () => {
+    const candidate = Object.assign({}, deepCandidate)
+    deepCandidate.style.color.name.name = 1 as any
+
+    const validationEither = validate(candidate, deepDescription, {
+      recurse: true
+    })
+
+    assert(isLeft(validationEither))
+    const error = unwrapEither(validationEither)
+    assert(error.code === 'INVALID_PROPERTIES')
+  })
 
 })